requirements.txt
aiohttp==3.6.2
aiopg==1.0.0
aioredis==1.3.1
python-socketio==4.4.0
psycopg2-binary==2.8.4
SQLAlchemy==1.3.13
